The code now compiles on JDK 1.8

+++ src/main/java/redis/clients/johm/collections/RedisList.java
@@ -262,7 +262,7 @@ public class RedisList<T> implements java.util.List<T> {
             if (johmElementType == JOhmCollectionDataType.PRIMITIVE) {
                 elements.add((T) Convertor.convert(elementClazz, key));
             } else if (johmElementType == JOhmCollectionDataType.MODEL) {
-                elements.add((T) JOhm.get(elementClazz, Integer.valueOf(key), ignoring));
+                elements.add((T) JOhm.get(elementClazz, Integer.valueOf(key).intValue(), ignoring));
             }
         }
         return elements;
